FEAT: implementado funcionalidade de login e retorno dos usuarios

+++ api/src/main/java/br/com/dferias/api/model/DTO/FuncionarioDTO.java
@@ -13,8 +13,8 @@ import lombok.Setter;
 @NoArgsConstructor
 public class FuncionarioDTO {
 
-    private int id;
-    private int idEquipe;
+    private Long id;
+    private Long idEquipe;
     private String nome;
     private Date dataAdmissao;
     private String email;
